Added Sound, fixed display, removed minor bugs.
Sound was added via winsound module, The issue of the game crashing due to bad handling of object clipping at display was fixed by changing the pixel location from the list by turning it into its modulo of 2048 so that the clipping works fine. The emulator currently plays pong fine.
The key layout was changed.